backend/requirements.txt
# Web Framework
fastapi==0.115.6
uvicorn[standard]==0.32.1

# Database
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10
asyncpg==0.30.0

# Celery + Redis
celery==5.4.0
redis==5.2.1
flower==2.0.1

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12

# Config
pydantic==2.10.3
email-validator==2.2.0
pydantic-settings==2.6.1

# File handling
python-magic-bin==0.4.14; sys_platform == "win32"
python-magic==0.4.27; sys_platform != "win32"
aiofiles==24.1.0

# Utilities
httpx==0.28.1

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1
factory-boy==3.3.1